refactor: better code organization and fix command palette
- move Docker command execution logic to DockerActions mixin
- extract refresh functionality to RefreshActions mixin
- add command palette support with dynamic action availability
  - fix broken command palette not showing commands
- move logging setup to dedicated utils module
- extract UI widgets (ErrorDisplay, StatusBar) to separate modules
- improve separation of concerns and code organization
