+++ dockerview/ui/widgets/__init__.py
@@ -1,0 +1,5 @@
+"""UI widgets for dockerview."""
+
+from .status import ErrorDisplay, StatusBar
+
+__all__ = ["ErrorDisplay", "StatusBar"]
