+++ dockerview/utils/__init__.py
@@ -1,0 +1,5 @@
+"""Utility modules for dockerview."""
+
+from .logging import setup_logging
+
+__all__ = ["setup_logging"]
